requirements.txt
Pillow
requests
openai
pywin32; sys_platform == 'win32'
python-docx
pymupdf
# layoutparser # Install the base layoutparser library with  
# layoutparser[paddledetection] # Install DL layout model toolkit 
# layoutparser[ocr] # Install OCR toolkit
# torch
# torchvision
# torchaudio
# transformers